requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.23.1
pandas==1.4.3
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
lxml==4.9.1
numpy==1.23.1
pandas==1.4.3
pandas-datareader==0.10.0
Pillow==9.1.1
plotly==5.9.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
urllib3==1.26.10
